stats_service/requirements.txt
grpcio~=1.71.0
grpcio-tools~=1.71.0
protobuf~=5.29.5
python-dotenv~=1.1.0
pytest~=8.3.5
kafka-python~=2.2.10
clickhouse-connect~=0.8.17